move zchash only functions
