+++ contracts/satoshi-bridge/src/bitcoin_utils/mod.rs
@@ -1,4 +1,5 @@
 pub mod contract_methods;
 pub mod psbt_wrapper;
+pub mod refund;
 pub mod transaction;
 pub mod types;
