+++ contracts/satoshi-bridge/src/bitcoin_utils/refund.rs
@@ -1,0 +1,30 @@
+use crate::bitcoin_utils::types::ChainSpecificData;
+use crate::env;
+use crate::psbt_wrapper::PsbtWrapper;
+use crate::Contract;
+use near_sdk::PromiseOrValue;
+
+impl Contract {
+    /// Execute an approved refund request (Bitcoin). Builds the refund PSBT
+    /// synchronously. The caller chooses `timelock_sec` (pass `0` to bypass).
+    /// `chain_specific_data` is unused on Bitcoin (kept for a uniform API with
+    /// the Zcash variant).
+    pub fn internal_execute_refund(
+        &mut self,
+        utxo_storage_key: String,
+        timelock_sec: u64,
+        _chain_specific_data: Option<ChainSpecificData>,
+    ) -> PromiseOrValue<bool> {
+        let refund_request = self.load_refund_request_for_execute(&utxo_storage_key, timelock_sec);
+        let (outpoint, deposit_output, refund_amount) =
+            self.refund_execution_inputs(&refund_request);
+        let refund_output = self.build_refund_output(&refund_request.refund_address, refund_amount);
+
+        let mut psbt = PsbtWrapper::new(vec![outpoint], vec![refund_output]);
+        psbt.set_input_utxo(vec![deposit_output]);
+
+        let caller = env::predecessor_account_id();
+        self.finalize_refund_with_psbt(caller, refund_request, psbt, refund_amount, utxo_storage_key);
+        PromiseOrValue::Value(true)
+    }
+}
