+++ contracts/satoshi-bridge/src/zcash_utils/mod.rs
@@ -2,5 +2,6 @@
 pub mod contract_methods;
 pub mod orchard_policy;
 pub mod psbt_wrapper;
+pub mod refund;
 pub mod transaction;
 pub mod types;
